+++ tests/TheChest.Core.Tests/Containers/ContainerTests/ContainerTests.contains_amount.cs
@@ -8,7 +8,7 @@ namespace TheChest.Core.Tests.Containers.ContainerTests
         [IgnoreIfValueType]
         public void ContainsAmount_NullItem_ThrowsArgumentNullException()
         {
-            var container = this.containerFactory.EmptyContainer();
+            var container = this.containerFactory.Empty();
             Assert.Throws<ArgumentNullException>(() => container.Contains(default!, 1));
         }
 
@@ -16,7 +16,7 @@ namespace TheChest.Core.Tests.Containers.ContainerTests
         [IgnoreIfReferenceType]
         public void ContainsAmount_DefaultValue_ReturnsFalseIfEmpty()
         {
-            var slot = this.containerFactory.EmptyContainer();
+            var slot = this.containerFactory.Empty();
             Assert.That(slot.Contains(default!, 1), Is.False);
         }
 
@@ -24,7 +24,7 @@ namespace TheChest.Core.Tests.Containers.ContainerTests
         [IgnoreIfReferenceType]
         public void ContainsAmount_DefaultValue_ReturnsTrueIfFull()
         {
-            var slot = this.containerFactory.FullContainer(10, default!);
+            var slot = this.containerFactory.Full(10, default!);
             Assert.That(slot.Contains(default!, 10), Is.True);
         }
 
@@ -33,7 +33,7 @@ namespace TheChest.Core.Tests.Containers.ContainerTests
         public void ContainsAmount_InvalidAmount_ThrowsArgumentOutOfRangeException(int amount)
         {
             var item = this.itemFactory.CreateDefault();
-            var container = this.containerFactory.EmptyContainer();
+            var container = this.containerFactory.Empty();
             Assert.Throws<ArgumentOutOfRangeException>(() => container.Contains(item, amount));
         }
     }