+++ tests/TheChest.Core.Tests/Containers/Factories/ContainerFactory.cs
@@ -1,5 +1,4 @@
-﻿using TheChest.Core.Containers;
-using TheChest.Core.Containers.Interfaces;
+﻿using TheChest.Core.Containers.Interfaces;
 using TheChest.Core.Slots.Interfaces;
 using TheChest.Core.Tests.Extensions;
 
@@ -15,7 +14,7 @@ namespace TheChest.Core.Tests.Containers.Factories
             this.slotFactory = slotFactory;
         }
 
-        public virtual IContainer<Item> EmptyContainer(int size = 20)
+        public virtual IContainer<Item> Empty(int size = 20)
         {
             var containerType = typeof(Container).GetContainerType(typeof(IContainer<Item>));
             var slotType = containerType.GetSlotTypeByConstructor<ISlot<Item>>();
@@ -23,7 +22,7 @@ namespace TheChest.Core.Tests.Containers.Factories
             var slots = slotType
                 .CreateSlots(
                     size: size,
-                    slotFactory: _ => slotFactory.EmptySlot(),
+                    factory: _ => slotFactory.EmptySlot(),
                     shuffle: true
                 );
 
@@ -34,7 +33,7 @@ namespace TheChest.Core.Tests.Containers.Factories
             return (IContainer<Item>)container!;
         }
 
-        public virtual IContainer<Item> FullContainer(int size, Item item)
+        public virtual IContainer<Item> Full(int size, Item item)
         {
             var containerType = typeof(Container).GetContainerType(typeof(IContainer<Item>));
             var slotType = containerType.GetSlotTypeByConstructor<ISlot<Item>>();
@@ -42,7 +41,7 @@ namespace TheChest.Core.Tests.Containers.Factories
             var slots = slotType
                 .CreateSlots(
                     size: size,
-                    slotFactory: _ => slotFactory.FullSlot(item),
+                    factory: _ => slotFactory.FullSlot(item),
                     shuffle: true
                 );
             var container = Activator.CreateInstance(
@@ -53,7 +52,7 @@ namespace TheChest.Core.Tests.Containers.Factories
             return (IContainer<Item>)container!;
         }
 
-        public virtual IContainer<Item> ShuffledItemContainer(int size, Item item)
+        public virtual IContainer<Item> WithItemShuffled(int size, Item item)
         {
             if (item is null)
                 throw new ArgumentException("Item cannot be null");
@@ -65,7 +64,7 @@ namespace TheChest.Core.Tests.Containers.Factories
             var slots = slotType
                 .CreateSlots(
                     size: size,
-                    slotFactory: 
+                    factory: 
                         i => i == randomIndex
                             ? slotFactory.FullSlot(item)
                             : slotFactory.EmptySlot(),
@@ -80,7 +79,7 @@ namespace TheChest.Core.Tests.Containers.Factories
             return (IContainer<Item>)container!;
         }
 
-        public virtual IContainer<Item> ShuffledItemsContainer(int size, params Item[] items)
+        public virtual IContainer<Item> WithItemsShuffled(int size, params Item[] items)
         {
             if (items.Length > size)
                 throw new ArgumentException($"Item amount ({items.Length}) cannot be bigger than the container size ({size})");
@@ -91,10 +90,7 @@ namespace TheChest.Core.Tests.Containers.Factories
             var slots = slotType
                 .CreateSlots(
                     size: size,
-                    slotFactory:
-                        i => i < items.Length
-                            ? slotFactory.FullSlot(items[i])
-                            : slotFactory.EmptySlot(),
+                    factory: i => i < items.Length ? slotFactory.FullSlot(items[i]) : slotFactory.EmptySlot(),
                     shuffle: true
                 );
 
@@ -103,6 +99,28 @@ namespace TheChest.Core.Tests.Containers.Factories
                 args: new object[1] { slots }
             );
 
+            return (IContainer<Item>)container!;
+        }
+
+        public virtual IContainer<Item> WithItems(int size, params Item[] items)
+        {
+            if (items.Length > size)
+                throw new ArgumentException($"Item amount ({items.Length}) cannot be bigger than the container size ({size})");
+            
+            var containerType = typeof(Container).GetContainerType(typeof(IContainer<Item>));
+            var slotType = containerType.GetSlotTypeByConstructor<ISlot<Item>>();
+
+            var slots = slotType
+                .CreateSlots(
+                    size: size,
+                    factory: i => i < items.Length ? slotFactory.FullSlot(items[i]) : slotFactory.EmptySlot(),
+                    shuffle: false
+                );
+
+            var container = Activator.CreateInstance(
+                type: containerType,
+                args: new object[1] { slots }
+            );
             return (IContainer<Item>)container!;
         }
     }