+++ tests/TheChest.Core.Tests/Containers/Interfaces/IContainer/IContainerTests.is_full.cs
@@ -5,7 +5,7 @@
         [Test]
         public void IsFull_EmptySlots_ReturnsFalse() 
         {
-            var container = this.containerFactory.EmptyContainer();
+            var container = this.containerFactory.Empty();
 
             Assert.That(container.IsFull, Is.False);
         }
@@ -14,10 +14,9 @@
         public void IsFull_OneEmptySlot_ReturnsFalse()
         {
             var randomSize = random.Next(MIN_SIZE_TEST, MAX_SIZE_TEST);
-            var container = this.containerFactory.ShuffledItemsContainer(
-                randomSize, 
-                this.itemFactory.CreateMany(randomSize - 1)
-            );
+            var items = this.itemFactory.CreateMany(randomSize - 1);
+
+            var container = this.containerFactory.WithItemsShuffled(randomSize, items);
 
             Assert.That(container.IsFull, Is.False);
         }
@@ -25,11 +24,10 @@
         [Test]
         public void IsFull_OneFullSlot_ReturnsFalse()
         {
+            var item = this.itemFactory.CreateDefault();
             var randomSize = random.Next(MIN_SIZE_TEST, MAX_SIZE_TEST);
-            var container = this.containerFactory.ShuffledItemContainer(
-                randomSize,
-                this.itemFactory.CreateDefault()
-            );
+
+            var container = this.containerFactory.WithItemShuffled(randomSize, item);
 
             Assert.That(container.IsFull, Is.False);
         }
@@ -37,11 +35,10 @@
         [Test]
         public void IsFull_AllSlotsFull_ReturnsTrue()
         {
+            var item = this.itemFactory.CreateDefault();
             var randomSize = random.Next(MIN_SIZE_TEST, MAX_SIZE_TEST);
-            var container = this.containerFactory.FullContainer(
-                randomSize,
-                this.itemFactory.CreateDefault()
-            );
+
+            var container = this.containerFactory.Full(randomSize, item);
 
             Assert.That(container.IsFull, Is.True);
         }