+++ tests/TheChest.Core.Tests/Containers/Interfaces/IStackContainer/IStackContainerTests.contains.cs
@@ -6,7 +6,7 @@
         public void Contains_EmptyContainer_ReturnsFalse()
         {
             var item = this.itemFactory.CreateDefault();
-            var container = this.containerFactory.EmptyContainer();
+            var container = this.containerFactory.Empty();
             Assert.That(container.Contains(item), Is.False);
         }
 
@@ -14,7 +14,7 @@
         public void Contains_AllItemsDifferentFromParam_ReturnsFalse()
         {
             var item = this.itemFactory.CreateDefault();
-            var container = this.containerFactory.FullContainer(10, 1, item);
+            var container = this.containerFactory.Full(10, 1, item);
 
             var paramItem = this.itemFactory.CreateRandom();
             Assert.That(container.Contains(paramItem), Is.False);
@@ -26,7 +26,7 @@
             var item = this.itemFactory.CreateDefault();
             var items = this.itemFactory.CreateManyRandom(9)
                 .Append(item).ToArray();
-            var container = this.containerFactory.ShuffledItemsContainer(10, 10, items);
+            var container = this.containerFactory.ShuffledItems(10, 10, items);
 
             var paramItem = this.itemFactory.CreateDefault();
             Assert.That(container.Contains(paramItem), Is.True);