+++ tests/TheChest.Core.Tests/Containers/StackContainerTests/StackContainerTests.contains_amount.cs
@@ -8,7 +8,7 @@ namespace TheChest.Core.Tests.Containers.StackContainerTests
         [IgnoreIfValueType]
         public void ContainsAmount_NullItem_ThrowsArgumentNullException()
         {
-            var container = this.containerFactory.EmptyContainer();
+            var container = this.containerFactory.Empty();
             Assert.Throws<ArgumentNullException>(() => container.Contains(default!, 1));
         }
 
@@ -16,7 +16,7 @@ namespace TheChest.Core.Tests.Containers.StackContainerTests
         [IgnoreIfReferenceType]
         public void ContainsAmount_DefaultValue_ReturnsFalseIfEmpty()
         {
-            var slot = this.containerFactory.EmptyContainer();
+            var slot = this.containerFactory.Empty();
             Assert.That(slot.Contains(default!, 1), Is.False);
         }
 
@@ -24,8 +24,24 @@ namespace TheChest.Core.Tests.Containers.StackContainerTests
         [IgnoreIfReferenceType]
         public void ContainsAmount_DefaultValue_ReturnsTrueIfFull()
         {
-            var slot = this.containerFactory.FullContainer(10, 10, default!);
+            var slot = this.containerFactory.Full(10, 10, default!);
             Assert.That(slot.Contains(default!, 10), Is.True);
+        }
+
+        [Test]
+        [IgnoreIfValueType]
+        public void ContainsAmount_NotEnoughItems_ReturnsFalse()
+        {
+            var size = this.random.Next(5, 20);
+            var searchItemAmount = this.random.Next(1, size - 1);
+            var stackSize = this.random.Next(5, 10);
+            var items = this.itemFactory.CreateMany(searchItemAmount)
+                .Concat(this.itemFactory.CreateManyRandom(size - searchItemAmount))
+                .ToArray();
+            var container = this.containerFactory.ShuffledItems(size, stackSize, items);
+
+            var item = this.itemFactory.CreateDefault();
+            Assert.That(container.Contains(item, (searchItemAmount * stackSize) + 1), Is.False);
         }
 
         [TestCase(0)]
@@ -33,7 +49,7 @@ namespace TheChest.Core.Tests.Containers.StackContainerTests
         public void ContainsAmount_InvalidAmount_ThrowsArgumentOutOfRangeException(int amount)
         {
             var item = this.itemFactory.CreateDefault();
-            var container = this.containerFactory.EmptyContainer();
+            var container = this.containerFactory.Empty();
             Assert.Throws<ArgumentOutOfRangeException>(() => container.Contains(item, amount));
         }
     }